Updated Name to Inventory from Shopping
