+++ Inventory.Extensions/Inventory.Extensions.csproj
