+++ Inventory.WebApp/Properties/launchSettings.json
