+++ Inventory.WebApp/StartupExtension.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Inventory.SqlDbProvider;
+
+namespace Inventory.WebApp
+{
+    public static class StartupExtension
+    {
+        public static void AddBindings(this IServiceCollection services, IConfiguration config)
+        {
+            services.AddDataProviders(config);
+        }
+    }
+}
