+++ Inventory.WebApp/appsettings.Development.json
