+++ Inventory.WebApp/appsettings.json
